+++ include/pcl_utils/pcl_feature_node.hpp
@@ -2,8 +2,8 @@
 // MIT License
 // https://github.com/yamaha-bps/ros2_pcl_utils/blob/master/LICENSE
 
-#ifndef PCL_UTILS__PCL_FEATURE_COMPONENT_HPP_
-#define PCL_UTILS__PCL_FEATURE_COMPONENT_HPP_
+#ifndef PCL_UTILS__PCL_FEATURE_NODE_HPP_
+#define PCL_UTILS__PCL_FEATURE_NODE_HPP_
 
 #include <rclcpp/rclcpp.hpp>
 #include <sensor_msgs/msg/point_cloud2.hpp>
@@ -15,11 +15,11 @@
 namespace cbr
 {
 
-class PclFeatureComponent : public rclcpp::Node
+class PclFeatureNode : public rclcpp::Node
 {
 public:
-  explicit PclFeatureComponent(const rclcpp::NodeOptions & opts);
-  ~PclFeatureComponent();
+  explicit PclFeatureNode(const rclcpp::NodeOptions & opts);
+  ~PclFeatureNode();
 
 private:
   rclcpp::Subscription<sensor_msgs::msg::PointCloud2>::SharedPtr sub_pcl_;
@@ -33,4 +33,4 @@ private:
 
 }  // namespace cbr
 
-#endif  // PCL_UTILS__PCL_FEATURE_COMPONENT_HPP_
+#endif  // PCL_UTILS__PCL_FEATURE_NODE_HPP_
